refactor(cli): move the cli code to the cli folder
